chore: 通知更新時間參數修正為Long

+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/notify/get/GetNotifyResponseBody.kt
@@ -42,7 +42,7 @@ data class GetNotifyResponseBody(
     @SerializedName("title")
     val title: String?,
     @SerializedName("updateTime")
-    val updateTime: Int?,
+    val updateTime: Long?,
     @SerializedName("new")
     val new : Any?
 )
